Overhaul of the entire site
This is a massive commit that introduces an entire new site.
This is what happens when you decide to code while you are stuck at
home during covid and don't have the enegry to manage a "wee" project.

Documentation and unit tests to come, but the code was written with
tdd in mind